refactor: move RobotState functionality into Drive subsystem and update vision integration
